The previous commit had a bug in Robot.java that threw an error that dissallowed the roboRio to execute the code. The error is fixed, but the cause is unknown. May have been initializing the methods outside of robotInit().
Implemented ramprate for starting (should do some testing at different
speeds). This still has to be implemented for stopping, if possible.
There is a separate ramprate for closed-loop systems (ie PID, part of
which we will be using), so will have to work on that when get encoders
working.

Updated the firmware of the 4 connected talons and of the PDP. Note for
future -- if something doesn't work like it should according to the API,
check if the firmware is up to date (the update enabled ramprate). The
update firmware can be found under "C:\Users\Public\Documents\FRC" on
Windows or equivalent. To update, go to the web interface, login with
username Admin and password field blank, choose the device to update,
the rest should be self-explanatory.

We have to wire the light to the RSL terminal on the RoboRio.

Will want to experiment with RobotBuilder. Our code may be fragmented in
the sence of it not following conventions for command based programming.